+++ src/demo.h
@@ -7,13 +7,13 @@
 class Demo {
   public:
     Demo(sync_device *rocket, MusicPlayer const &player);
-    float get(std::string track);
+    double get(std::string track);
     void render();
 
   private:
     sync_device *rocket;
     MusicPlayer const &player;
-    float t;
+    double t;
     // Likely faster to cache tracks to a map as sync_get_track only iterates.
     std::unordered_map<std::string, sync_track const *> tracks;
 };
